Renaming to match maturity -> availability and compliance -> certification
